requirements.txt
pandas==2.0.2
Pillow==9.5.0
Pillow==10.1.0
plotly==5.15.0
streamlit==1.23.1
